Split swift helpers in SwiftUI and Swift helpers
